+++ app/assets/scripts/app.js
@@ -1,7 +1,9 @@
 define( ['angular','ngAnimate','minireel/app','account/app','login','portal','c6uilib','c6log',
-         'c6_defines','templates','forgot_password','ui','version','selfie','selfie/app'],
+         'c6_defines','templates','forgot_password','ui','version','selfie','selfie/app',
+         'libs'],
 function( angular , ngAnimate , minireel     , account     , login , portal , c6uilib , c6log ,
-          c6Defines  , templates , forgotPassword  , ui , version , selfie , selfieApp ) {
+          c6Defines  , templates , forgotPassword  , ui , version , selfie , selfieApp,
+          libs ) {
     'use strict';
 
     var forEach = angular.forEach,
@@ -33,7 +35,8 @@ function( angular , ngAnimate , minireel     , account     , login , portal , c6
         portal.name,
         forgotPassword.name,
         selfie.name,
-        selfieApp.name
+        selfieApp.name,
+        libs.name
     ])
         .config(['c6UrlMakerProvider',
         function( c6UrlMakerProvider ) {
@@ -1327,7 +1330,7 @@ function( angular , ngAnimate , minireel     , account     , login , portal , c6
                     'https://fonts.googleapis.com/css?family=Source+Sans+Pro:300,400,' +
                         '600,700,300italic,400italic,600italic,700italic|Roboto+Conde' +
                         'nsed:300italic,400italic,300,400',
-                    'https://maxcdn.bootstrapcdn.com/font-awesome/4.3.0/css/font-awesome.min.css',
+                    'https://maxcdn.bootstrapcdn.com/font-awesome/4.5.0/css/font-awesome.min.css',
                     'styles/selfie/css/select2.min.css',
                     'styles/selfie/css/css-wizardry-grids.css',
                     'styles/selfie/css/c6selfie__base.css',