+++ test/test-main.js
@@ -59,6 +59,7 @@
                 c6uilib: libUrl('c6ui/v3.7.4-0-g734d5f3/c6uilib'),
                 c6log: libUrl('c6ui/v3.7.4-0-g734d5f3/c6log'),
                 c6embed: libUrl('c6embed/v1/app'),
+                metagetta: libUrl('metagetta/v0.2.0-0-g8b51280/metagetta'),
                 braintree: 'lib/braintree',
                 'helpers/drag': '/base/test/helpers/drag'
             },