+++ CHANGELOG.md
@@ -1,13 +1,16 @@
 # MiniReelinator
 
 * *[v1.28.1.rc1]*
+  * Show warning in the admin tab if there is no payment method
+  * [FEATURE]: Show start and end dates in the admin tab
+  * [FEATURE]: Add support for Instagram video-only cards
   * [DESIGN]: Added styles for campaign confirmation modal
   * [FEATURE]: Add Flight Dates in campaign creator/editor view
   * [DESIGN]: Added styles for campaign stats screen
   * [FEATURE]: Add Campaign Summary modal when submitting
   * [FIX]: Update CPV calculations
   * [FIX]: Show warning when choosing flight dates within the next two days
-  * [DESIGN]: Added styles for datepickers to fix IE bug and to higlight selected dates
+  * [DESIGN]: Added styles for icons on stats page, updated font-awesome library
 
 ## v1.28.0 (December 1, 2015)
 * *[v1.28.0.rc1]*